refactor: seperate infra, service, api
