+++ pdf_server/consts.py
@@ -1,0 +1,9 @@
+import mimetypes
+from functools import lru_cache
+
+mimetypes.init()
+
+@lru_cache
+def get_mimetype(ext: str):
+    """etx example .doc .docx"""
+    return mimetypes.types_map.get(ext)
